Update field type in SolicitorCosts class to allow decimal values.
This change will align the class with the MAAT equivalent data structure, which does allow hours to be input as a decimal.

+++ crime-hardship/src/main/java/uk/gov/justice/laa/crime/hardship/dto/maat_api/SolicitorCosts.java
@@ -13,7 +13,7 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class SolicitorCosts {
     private BigDecimal rate;
-    private Integer hours;
+    private BigDecimal hours;
     private BigDecimal vat;
     private BigDecimal disbursements;
     private BigDecimal estimatedTotal;
